requirements.txt
beautifulsoup4==4.12.2
canal-python==0.4
confluent-kafka==2.3.0
google==3.0.0
kazoo==2.5.0
mysql-replication==1.0.5
packaging==23.2
protobuf==3.20.1
pykafka==2.8.0
PyMySQL==1.1.0
six==1.16.0
soupsieve==2.5
tabulate==0.9.0
